requirements.txt

fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pillow==10.0.1
numpy==1.24.4
pydantic==2.4.2
python-dotenv==1.0.0
aiofiles==23.2.1
